# Conflicts:
#	app/src/main/java/com/gox/partner/fcm/FcmService.kt
#	app/src/main/res/layout/bank_template_inflate.xml
#	base/build.gradle
#	build.gradle
#	taxiservice/src/main/java/com/gox/taxiservice/views/main/TaxiDashboardActivity.kt

+++ app/src/main/java/com/gox/partner/views/sign_in/LoginActivity.kt
@@ -115,7 +115,7 @@ class LoginActivity : BaseActivity<ActivityLoginBinding>(), LoginViewModel.Login
 
 
     fun setDefaultCountry(){
-        val dr = ContextCompat.getDrawable(this, R.drawable.flag_india as Int)
+        val dr = ContextCompat.getDrawable(this, R.drawable.flag_india)
         val bitmap = (dr as BitmapDrawable).bitmap
         var width:Int=0
         var height:Int=0
